+++ frontend/src/metabase/query_builder/VisualizationError.jsx
@@ -1,5 +1,4 @@
 import React, { Component, PropTypes } from 'react';
-import { findDOMNode } from 'react-dom';
 
 import MetabaseSettings from "metabase/lib/settings";
 import VisualizationErrorMessage from './VisualizationErrorMessage';
@@ -16,15 +15,17 @@ const EmailAdmin = () => {
 }
 
 class VisualizationError extends Component {
-  constructor (props) {
-      super(props)
-  }
 
-  showDetailError() {
-      if (this._detailErrorLink && this._detailErrorBody ) {
-          findDOMNode(this._detailErrorLink).style.display = "none";
-          findDOMNode(this._detailErrorBody).style.display = "inherit";
+  constructor(props) {
+      super(props);
+      this.state = {
+          showError: false
       }
+  }
+  static propTypes = {
+      card:     PropTypes.object.isRequired,
+      duration: PropTypes.number.isRequired,
+      error:    PropTypes.object.isRequired,
   }
 
   render () {
@@ -66,10 +67,10 @@ class VisualizationError extends Component {
                   <div className="QueryError2-details">
                       <h1 className="text-bold">There was a problem with your question</h1>
                       <p className="QueryError-messageText">Most of the time this is caused by an invalid selection or bad input value.  Double check your inputs and retry your query.</p>
-                      <div ref={(c) => this._detailErrorLink = c} className="pt2">
-                          <a onClick={this.showDetailError.bind(this)} className="link cursor-pointer">Show error details</a>
+                      <div className="pt2">
+                          <a onClick={() => this.setState({ showError: true })} className="link cursor-pointer">Show error details</a>
                       </div>
-                      <div ref={(c) => this._detailErrorBody = c} style={{display: "none"}} className="pt3 text-left">
+                      <div style={{ display: this.state.showError? 'inherit': 'none'}} className="pt3 text-left">
                           <h2>Here's the full error message</h2>
                           <div style={{fontFamily: "monospace"}} className="QueryError2-detailBody bordered rounded bg-grey-0 text-bold p2 mt1">{error}</div>
                       </div>
